Add lots of required devices and drivers : be carefull, lots of code is inactivated
- Coders : CoderErrorComputer.c, trajectory.c
- Config : robotConfigDevice.c,
- StartMatchDetectorDevice
- strategy.c : LCD removed

+++ drivers/beacon/beaconReceiverDriver.c
@@ -26,10 +26,10 @@ bool updateOpponentRobotPosition(Point* opponentRobotPosition) {
 
     bool result = transmitFromDriverRequestBuffer();
 
-    opponentRobotPosition->x = readHex4(inputStream);
+    opponentRobotPosition->x = (float) readHex4(inputStream);
     // separator    
     readHex(inputStream);
-    opponentRobotPosition->y = readHex4(inputStream);
+	opponentRobotPosition->y = (float) readHex4(inputStream);
 
     return result;
 }
